factor out feed and links mixins for opds 1
